Fix for Linux

+++ src/map/utils/petutils.cpp
@@ -1920,7 +1920,7 @@ namespace petutils
 
             // And cap it to the master's level or their item's level if the player is level 99
             auto capLevel = 0;
-            if ((lua["xi"]["settings"]["main"]["MAX_LEVEL"].get<uint8>() >= 99) && (PMaster->GetMLevel() >= 99))
+            if ((lua["xi"]["settings"]["main"]["MAX_LEVEL"].get<uint16>() >= 99) && (PMaster->GetMLevel() >= 99))
             {
                 capLevel = std::max(PMaster->GetMLevel(), PMaster->m_Weapons[SLOT_MAIN]->getILvl());
             }
